Extract import and export parsing code
