+++ src/parser.common.ts
@@ -1,3 +1,7 @@
+import * as ts from 'typescript';
+
+export const star = ['*'];
+
 export interface FromWhat {
   from: string;
   what: string[];
@@ -7,3 +11,6 @@ export const TRIM_QUOTES = /^['"](.*)['"]$/;
 
 export const getFromText = (moduleSpecifier: string): string =>
   moduleSpecifier.replace(TRIM_QUOTES, '$1').replace(/\/index$/, '');
+
+export const getFrom = (moduleSpecifier: ts.Expression): string =>
+  getFromText(moduleSpecifier.getText());
